fixes for magic modem automation scripts (for uncommon paths mostly) + temporary changes (awaiting fixes) + CONFIG fixes

+++ oldmodem.sh
@@ -4,27 +4,23 @@ make mrproper
 
 if test -s Makefile_backup -a -s arch/arm/Makefile_backup -a -d drivers/misc/samsung_modemctl_backup; then
 
-	mv Makefile Makefile_opti
-	mv Makefile_backup Makefile
-
-	cd arch/arm && {
 		mv Makefile Makefile_opti
 		mv Makefile_backup Makefile
-	}
 
-	cd ~/android/system/kernel/samsung/glitch/drivers/misc && {
-		mv samsung_modemctl samsung_modemctl_opti
-		mv samsung_modemctl_backup samsung_modemctl
-	}
+		mv arch/arm/Makefile arch/arm/Makefile_opti
+		mv arch/arm/Makefile_backup arch/arm/Makefile
 
-	echo "Switching done. Building..."
+		mv drivers/misc/samsung_modemctl drivers/misc/samsung_modemctl_opti
+		mv drivers/misc/samsung_modemctl_backup drivers/misc/samsung_modemctl
+
+	echo "Switching done. Building..." && {
 	
-	cd ~/android/system/kernel/samsung/glitch && {
 	./galaxys.sh
 	}
+
 	#Delete the created old-toolchain kernel for less confusion
 	
-	cd ~/android/system/kernel/samsung/glitch/release/SGS && {
+	cd release/SGS && {
 	declare -i RELVER=0
 
 	until test ! -s ${REL}; do
@@ -35,38 +31,49 @@ if test -s Makefile_backup -a -s arch/arm/Makefile_backup -a -d drivers/misc/sam
 	
 	rm ${REL_}
 	rm ${REL_}.sha256sum
+	cd ../../
 	}
 	#Deleted	
 	
-	echo "Building done. Reswitching..."
+	echo "Building done. Reswitching..." && {
 
-	cd ~/android/system/kernel/samsung/glitch && {
 		mv Makefile Makefile_backup
 		mv Makefile_opti Makefile
-	}
 
-	cd arch/arm && {
-		mv Makefile Makefile_backup
-		mv Makefile_opti Makefile
-	}
+		mv arch/arm/Makefile arch/arm/Makefile_backup
+		mv arch/arm/Makefile_opti arch/arm/Makefile
 
-	cd ~/android/system/kernel/samsung/glitch/drivers/misc && {
+		mv drivers/misc/samsung_modemctl/built-in.o drivers/misc/samsung_modemctl_opti/built-in.443stock_samsung_modemctl
+		mv drivers/misc/samsung_modemctl/modemctl/built-in.o drivers/misc/samsung_modemctl_opti/modemctl/built-in.443stock_modemctl
 
-		mv samsung_modemctl/built-in.o samsung_modemctl_opti/built-in.443stock_samsung_modemctl
-		mv samsung_modemctl/modemctl/built-in.o samsung_modemctl_opti/modemctl/build-in.443stock_modemctl
+		mv drivers/misc/samsung_modemctl drivers/misc/samsung_modemctl_backup
+		mv drivers/misc/samsung_modemctl_opti drivers/misc/samsung_modemctl
+		}
 
-		mv samsung_modemctl samsung_modemctl_backup
-		mv samsung_modemctl_opti samsung_modemctl
-	}
+	echo "Done! now preparing for next build..." && {
 
-	echo "Done! now preparing for next build..."
-
-	cd ~/android/system/kernel/samsung/glitch
 	./prepare.sh
+	}
 
 else
 
-	echo "There's a backup missing"
+	echo "There's a backup missing, if needed, reverting changes to prepare rebuild "
+
+fi
+
+if test -s Makefile_opti -a -s arch/arm/Makefile_opti -a -d drivers/misc/samsung_modemctl_opti; then
+
+		mv Makefile Makefile_backup
+		mv Makefile_opti Makefile 
+
+		mv arch/arm/Makefile arch/arm/Makefile_backup
+		mv arch/arm/Makefile_opti arch/arm/Makefile
+
+		mv drivers/misc/samsung_modemctl drivers/misc/samsung_modemctl_backup
+		mv drivers/misc/samsung_modemctl_opti drivers/misc/samsung_modemctl
+
+	echo "Done!"
+
 
 fi
